+++ PHEngine/PHEngine/src/Core/GameCore/Physics/PhysicsDescriptors/Shapes/PhyPlaneShape.h
@@ -10,7 +10,7 @@ namespace EnginePhysics
    struct PhyPlaneShape
       : public PhysicsShapeBase
    {
-      PhyPlaneShape(const glm::vec3& normal, const float d)
+      PhyPlaneShape(const glm::vec3& normal, const double d)
          : PhysicsShapeBase(new btStaticPlaneShape(btVector3(normal.x, normal.y, normal.z), d))
       {
       }
